[Effects] Remove wrapping by moving effect logic into new interface implementation classes.
